+++ lib/oli_web/router.ex
@@ -766,16 +766,19 @@ defmodule OliWeb.Router do
     resources("/open_and_free", OpenAndFreeController, as: :admin_open_and_free)
     live("/open_and_free/:section_slug/remix", Delivery.RemixSection, as: :open_and_free_remix)
 
-    # Publishing Institutions and Registrations
+    # Institutions, LTI Registrations and Deployments
     resources("/institutions", InstitutionController)
 
-    resources("/registrations", RegistrationController) do
+    live("/registrations", Admin.RegistrationsView)
+
+    resources("/registrations", RegistrationController, except: [:index]) do
       resources("/deployments", DeploymentController, except: [:index, :show])
     end
 
     put("/approve_registration", InstitutionController, :approve_registration)
     delete("/pending_registration/:id", InstitutionController, :remove_registration)
 
+    # Communities
     live("/communities/new", CommunityLive.NewView)
 
     # Course Ingestion